Included Decision Tree model
Updated the machine learning file with Decision Tree model and removed models that have been implemented from rubric. Also I just realized that the comparison between the GUI and ml.py file is case sensitive. Might or might not be fixed in the future. To be implemented: Conversion of strings into numerical